+++ Assembler/token.h
@@ -6,7 +6,7 @@
 struct Token {
     union {
         const char *text;
-        int16_t value;
+        int32_t value;
     };
     uint16_t type;
 };
@@ -24,7 +24,7 @@ struct Token {
 #define RPAREN 10
 
 struct Token newTokenT(const char *text, uint16_t type);        // Create new object
-struct Token newTokenV(int16_t value, uint16_t type);   // Create new object
+struct Token newTokenV(int32_t value, uint16_t type);   // Create new object
 char *tokenToStr(struct Token token);   // Convert to string
 const char *typeName(uint16_t type);
